backend/requirements.txt
fastapi
uvicorn
python-multipart
pinecone-client
langchain
langchain-pinecone
langchain-huggingface
sentence-transformers
pdfplumber
pydantic
